Script to test COS value using Scapy
This script will initiate traffic to a Cisco IOX-XR and receive traffic with a COS=5 (marked by IOS-XR). Then at the ingress interface it will open all the received packets through the Scapy module and verify the COS value has been assigned correctly to each packet. If there is an anomaly, then it will report.